+++ management-api-server/src/main/java/com/datastax/mgmtapi/resources/TableOpsResources.java
@@ -217,7 +217,7 @@ public class TableOpsResources extends BaseResources {
                 "CALL NodeOps.forceKeyspaceCompaction(?, ?, ?, ?)",
                 compactRequest.splitOutput,
                 keyspaceName,
-                compactRequest.tables,
+                tables,
                 false);
           }